+++ src/BuildEngineFactory.ts
@@ -3,9 +3,9 @@
  * Copyright (c) 2020 Handsoncode. All rights reserved.
  */
 
+import {Exporter} from './Exporter'
 import {LazilyTransformingAstService} from './LazilyTransformingAstService'
 import {CellContentParser} from './CellContentParser'
-import {Exporter} from './CellValue'
 import {buildColumnSearchStrategy, ColumnSearchStrategy} from './Lookup/SearchStrategy'
 import {Config, ConfigParams} from './Config'
 import {DateTimeHelper} from './DateTimeHelper'
