+++ src/HyperFormula.ts
@@ -6,7 +6,8 @@
 import {AbsoluteCellRange} from './AbsoluteCellRange'
 import {CellType, CellValueType, getCellType, getCellValueType, SimpleCellAddress} from './Cell'
 import {CellContent, CellContentParser, RawCellContent} from './CellContentParser'
-import {CellValue, ExportedChange, Exporter, NoErrorCellValue} from './CellValue'
+import {CellValue, NoErrorCellValue} from './CellValue'
+import {ExportedChange, Exporter} from './Exporter'
 import {ColumnSearchStrategy} from './Lookup/SearchStrategy'
 import {Config, ConfigParams} from './Config'
 import {ColumnRowIndex, CrudOperations} from './CrudOperations'
